requirementes.txt
djando
djangorestframework
pillow
psycopg2-binary
python-decouple
dj_rest_auth
